added details view page.
you have to manually got to localhost:5000 for some reason, but when the 
application is built it wont be a problem.

we probably need to have the page be able to be on react, so the user 
can easily navigate to and from their search results and the view and 
then when they are ready to share the link they can. people who visit 
from third party wont be able to view the original search box, unless we 
add it somewhere in the url. right now each post is shoved in the url.